requirements.txt
tornado
pgmagic
